+++ src/StateSmith/Input/PlantUML/PlantUMLWalker.cs
@@ -11,16 +11,16 @@ namespace StateSmith.Input.PlantUML
         public List<DiagramEdge> edges = new();
         public DiagramNode root = new();
 
-        private DiagramNode currentState;
+        private DiagramNode currentNode;
 
-        private Dictionary<DiagramNode, DiagramNode> stateInitalStateMap = new();
-
+        private Dictionary<DiagramNode, DiagramNode> nodeInitalStateMap = new();
+        public Dictionary<DiagramNode, string> nodeStereoTypeLookup = new();
         public Dictionary<string, DiagramNode> nodeMap = new();
 
 
         public PlantUMLWalker()
         {
-            currentState = root;
+            currentNode = root;
         }
 
         public override void EnterState_explicit([NotNull] PlantUMLParser.State_explicitContext context)
@@ -31,44 +31,45 @@ namespace StateSmith.Input.PlantUML
 
             if (context.STRING() != null)
             {
-                label = context.STRING().GetText();
+                label = Decode(context.STRING().GetText());
                 label = label.Substring(1, label.Length - 2);   // remove quotes
             }
 
-            var state = GetOrAddState(id);
-            currentState = state;
-            state.id = id;
-            state.label = label;
+            var node = GetOrAddNode(id);
+            HandleStereotype(node, context.stereotype());
+            currentNode = node;
+            node.id = id;
+            node.label = label;
         }
 
         public override void ExitState_explicit([NotNull] PlantUMLParser.State_explicitContext context)
         {
-            currentState = currentState.parent;
+            currentNode = currentNode.parent;
         }
 
         private DiagramNode GetOrAddInitialState(PlantUMLParser.VertexContext vertexContext)
         {
-            if (!stateInitalStateMap.TryGetValue(currentState, out var initialState))
+            if (!nodeInitalStateMap.TryGetValue(currentNode, out var initialState))
             {
                 initialState = new DiagramNode
                 {
                     label = Compiling.Compiler.InitialStateString,
-                    id = vertexContext.start_end_state().Start.Line + ":" + vertexContext.start_end_state().Start.Column
+                    id = vertexContext.start_end_state().Start.Line + "_" + vertexContext.start_end_state().Start.Column
                 };
-                stateInitalStateMap.Add(currentState, initialState);
-                AddState(initialState);
+                nodeInitalStateMap.Add(currentNode, initialState);
+                AddNode(initialState);
             }
 
             return initialState;
         }
 
-        private void AddState(DiagramNode state)
+        private void AddNode(DiagramNode state)
         {
-            currentState.children.Add(state);
-            state.parent = currentState;
+            currentNode.children.Add(state);
+            state.parent = currentNode;
         }
 
-        private DiagramNode GetOrAddState(string id)
+        private DiagramNode GetOrAddNode(string id)
         {
             if (!nodeMap.TryGetValue(id, out var state))
             {
@@ -78,7 +79,7 @@ namespace StateSmith.Input.PlantUML
                     id = id
                 };
                 nodeMap.Add(id, state);
-                AddState(state);
+                AddNode(state);
             }
 
             return state;
@@ -87,13 +88,14 @@ namespace StateSmith.Input.PlantUML
         public override void EnterDiagram([NotNull] PlantUMLParser.DiagramContext context)
         {
             root.id = context.startuml().identifier()?.GetText() ?? "ROOT";
+            root.label = "$STATEMACHINE : " + root.id;
         }
 
         public override void EnterTransition([NotNull] PlantUMLParser.TransitionContext context)
         {
             DiagramNode source = VertexToNode(context.vertex(0));
             DiagramNode destination = VertexToNode(context.vertex(1));
-            var label = context.transition_event_guard_code()?.GetText().Trim() ?? "";
+            var label = Decode(context.transition_event_guard_code()?.GetText());
 
             var edge = new DiagramEdge
             {
@@ -113,27 +115,63 @@ namespace StateSmith.Input.PlantUML
 
         public override void EnterState_contents([NotNull] PlantUMLParser.State_contentsContext context)
         {
-            var state = GetOrAddState(context.identifier().GetText());
-            state.label += "\n" + context.rest_of_line().GetText().Trim();
+            var state = GetOrAddNode(context.identifier().GetText());
+            state.label += "\n" + Decode(context.rest_of_line().GetText());
+        }
+
+        private string Decode(string? escapedString)
+        {
+            if (escapedString == null)
+            {
+                return "";
+            }
+
+            return escapedString.Trim().Replace(@"\n", "\n");
         }
 
         private static void ThrowValidationFailure(string message, PlantUMLParser.TransitionContext context)
         {
             int line = context.Start.Line;
             int column = context.Start.Column;
-            throw new ArgumentException($"{message}. Location Details {{ line: {line}, column: {column}, text: `{context.GetText()}`. }}");
+            throw new InvalidOperationException($"{message}. Location Details {{ line: {line}, column: {column}, text: `{context.GetText()}`. }}"); //todolow throw a custom exception type
         }
 
         private DiagramNode VertexToNode(PlantUMLParser.VertexContext vertexContext)
         {
+            DiagramNode node;
+
             if (vertexContext.start_end_state() != null)
             {
-                return GetOrAddInitialState(vertexContext);
+                node = GetOrAddInitialState(vertexContext);
             }
             else
             {
-                return GetOrAddState(vertexContext.state_id().GetText());
+                node = GetOrAddNode(vertexContext.state_id().GetText());
             }
+
+            HandleStereotype(node, vertexContext.stereotype());
+
+            return node;
+        }
+
+        private void HandleStereotype(DiagramNode node, PlantUMLParser.StereotypeContext stereotypeContext)
+        {
+            if (stereotypeContext == null)
+            {
+                return;
+            }
+
+            string newStereotype = stereotypeContext.identifier().GetText();
+
+            if (nodeStereoTypeLookup.TryGetValue(node, out var existingStereotype))
+            {
+                if (existingStereotype != newStereotype)
+                {
+                    throw new InvalidOperationException($"node stereotype redefined from `{existingStereotype}` to `{newStereotype}` ");  //todolow throw a custom exception type
+                }
+            }
+
+            nodeStereoTypeLookup.Add(node, newStereotype);
         }
     }