+++ src/StateSmith/compiler/Compiler.cs
@@ -21,15 +21,22 @@ namespace StateSmith.Compiling
         private List<string> eventNames = new List<string>();
         private Dictionary<Input.DiagramNode, Vertex> diagramVertexMap = new Dictionary<Input.DiagramNode, Vertex>();
 
-        public void CompileFile(string filepath)
+        /// <summary>
+        /// If you want to support an input source other than a yEd file, see <see cref="CompileDiagramNodesEdges(List{DiagramNode}, List{DiagramEdge})"/> instead.
+        /// </summary>
+        /// <param name="filepath"></param>
+        public void CompileYedFile(string filepath)
         {
             YedParser yedParser = new YedParser();
-
             yedParser.Parse(filepath);
-
             CompileDiagramNodesEdges(yedParser.GetRootNodes(), yedParser.GetEdges());
         }
 
+        /// <summary>
+        /// Call this method when you want to support a custom input source.
+        /// </summary>
+        /// <param name="rootNodes"></param>
+        /// <param name="edges"></param>
         public void CompileDiagramNodesEdges(List<DiagramNode> rootNodes, List<DiagramEdge> edges)
         {
             foreach (var node in rootNodes)
@@ -238,7 +245,7 @@ namespace StateSmith.Compiling
         {
             if (labelParser.HasError())
             {
-                throw new DiagramEdgeParseException(edge, sourceVertex, targetVertex, ParserErrorsToReasonStrings(labelParser, "\n"));
+                throw new DiagramEdgeParseException(edge, sourceVertex, targetVertex, ParserErrorsToReasonStrings(labelParser.GetErrors(), "\n"));
             }
         }
 
@@ -407,7 +414,7 @@ namespace StateSmith.Compiling
         {
             if (labelParser.HasError())
             {
-                string reasons = ParserErrorsToReasonStrings(labelParser, separator: "\n           ");
+                string reasons = ParserErrorsToReasonStrings(labelParser.GetErrors(), separator: "\n           ");
 
                 string parentPath = VertexPathDescriber.Describe(parentVertex);
                 string fullMessage = $@"Failed parsing node label
@@ -420,11 +427,12 @@ Reason(s): {reasons}
             }
         }
 
-        private static string ParserErrorsToReasonStrings(LabelParser labelParser, string separator)
+        // todolow move out of compiler
+        public static string ParserErrorsToReasonStrings(List<Error> errors, string separator)
         {
             var reasons = "";
             var needsSeparator = false;
-            foreach (var error in labelParser.GetErrors())
+            foreach (var error in errors)
             {
                 if (needsSeparator)
                 {