+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## WIP
+### Added
+- Support PlantUML input as an alternative to yEd. Input file extension must be one of ".pu", ".puml", ".plantuml"
+  https://github.com/StateSmith/StateSmith/issues/21
+
 ## [0.5.5-alpha]
 ### Fixed
 - Fix generated comment regarding marking event as handled when not for a transition